desktop/requirements.txt
flatten_dict
customtkinter==5.2.2
dash==2.14.2
Flask==3.0.0
future==0.18.3
CTKMessagebox==2.5
pysam==0.22.0
numpy==1.26.3
pandas==2.1.4
Pillow==10.2.0
plotly==5.12.0
requests==2.31.0
scipy==1.11.4
tkcalendar==1.6.1
tqdm==4.65.0
setuptools==68.2.2
matplotlib==3.8.2
dash_bootstrap_components
biopython
